1. -修改新特性界面- 	2. -微调欧动集市价格显示-

+++ ODApp/3-ODBazaar/Model/ODBazaarExchangeSkillModel.h
@@ -74,7 +74,7 @@ typedef enum : NSUInteger
 
 @property (nonatomic, copy) NSString *title;
 
-@property (nonatomic, assign) int price;
+@property (nonatomic, assign) CGFloat price;
 
 @property (nonatomic, assign) int swap_type;
